[Ui]: Apply responsive design to pages 'lobby', 'rules', 'group'. [Assets]: Add mission members count images.
