+++ imports/ui/components/group_table.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { insert } from '../../api/groups/methods.js';
 import GroupRow from './group_row.jsx';
+import { MediumAndSmallDevice } from '../layouts/devices.jsx';
 
 export default class GroupTable extends React.Component {
   constructor(props) {
@@ -17,42 +18,40 @@ export default class GroupTable extends React.Component {
         <br/>
         <form data-parsley-validate className="form-horizontal form-label-left" onSubmit={this.createGroup}>
           <div className="form-group">
-            <label className="control-label col-md-3 col-sm-3 col-xs-12" htmlFor="first-name">Group Name <span className="required">*</span>
+            <label className="control-label col-sm-3" htmlFor="first-name">Group Name <span className="required">*</span>
             </label>
-            <div className="col-md-6 col-sm-6 col-xs-12">
-              <input type="text" ref="name" name="name" required="required" className="form-control col-md-7 col-xs-12"/>
+            <div className="col-sm-6">
+              <input type="text" ref="name" name="name" required="required" className="form-control"/>
             </div>
-            <button type="submit" className="btn btn-default">Create new group</button>
+            <div className="col-sm-3">
+              <button type="submit" className="btn btn-default">Create new group</button>
+            </div>
           </div>
         </form>
       </div>
     ) : null;
     return (
-      <div className="row">
-        <div className="col-md-12">
-          <div className="x_panel">
-            {formCreateGroup}
-            <div className="x_title">
-              <h2>Groups</h2>
-              <div className="clearfix"></div>
-            </div>
-            <div className="x_content">
-              <p>Simple table with group listing with players, situation and options</p>
-              <table className="table table-bordered projects">
-                <thead>
-                  <tr>
-                    <th style={{ width: '20%' }}>Group</th>
-                    <th>Players</th>
-                    <th>Situation</th>
-                    <th style={{ width: '20%' }}></th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {groups.map(g => <GroupRow group={g} key={g._id}/>)}
-                </tbody>
-              </table>
-            </div>
-          </div>
+      <div className="x_panel">
+        {formCreateGroup}
+        <div className="x_title">
+          <h2>Groups</h2>
+          <div className="clearfix"></div>
+        </div>
+        <div className="x_content">
+          <p>Simple table with group listing with players, situation and options</p>
+          <table className="table table-bordered projects">
+            <thead>
+              <tr>
+                <th>Group</th>
+                <MediumAndSmallDevice><th>Players</th></MediumAndSmallDevice>
+                <th>Situation</th>
+                <th></th>
+              </tr>
+            </thead>
+            <tbody>
+              {groups.map(g => <GroupRow group={g} key={g._id}/>)}
+            </tbody>
+          </table>
         </div>
       </div>
     );
